# packages/api/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Etimo Achievements",
    "version": "1.0.0",
    "license": {
      "name": "Swagger spec",
      "url": "/swagger.json"
    },
    "contact": {
      "name": "Etimo AB",
      "url": "https://etimo.se"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "http://45.77.89.90:4000",
      "description": "Etimo server"
    },
    {
      "url": "https://etimo-achievements-staging.herokuapp.com",
      "description": "Staging server"
    },
    {
      "url": "https://etimo-achievements.herokuapp.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyParameter": {
        "type": "apiKey",
        "in": "query",
        "name": "apiKey"
      }
    },
    "schemas": {
      "AccessToken": {
        "title": "AccessToken",
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Bearer access token used for authorization.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiIzZmRmMGNhNi1mNWI5LTRlMTUtYWFlOC04MWM0NTRmZGZmYjAiLCJzdWIiOiIzZDQxYjVkNC0zMmZhLTQxZDYtOTM5OS01NTFmOThmNWIxMjAiLCJuYW1lIjoiTmljbGFzIExpbmRzdGVkdCIsImVtYWlsIjoibmljbGFzLmxpbmRzdGVkdEBldGltby5zZSIsImlzcyI6ImV0aW1vLWFjaGlldmVtZW50cyIsImF1ZCI6ImV0aW1vLWFjaGlldmVtZW50cyIsImV4cCI6MTY0NzY4OTA5NywiaWF0IjoxNjQ3Njg1NDk3fQ.npzbTJxvdHYF5qYPbAq8cjWReDiQkGbEIEiMCVWNkic"
          },
          "token_type": {
            "type": "string",
            "description": "The type of token issued.",
            "example": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "In how many seconds the token will expire.",
            "example": 3600
          },
          "refresh_token": {
            "type": "string",
            "description": "The refresh token that can be used to get a new access token.",
            "example": "V2R3ARQDwpzrLPXhsBZgexPMCp7mPQfsnATPie6h5RCdmmkU4UAqKfWvypPkhbHG"
          }
        }
      },
      "Achievement": {
        "title": "Achievement",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "The name of the achievement",
            "example": "Coffee Achievement"
          },
          "description": {
            "type": "string",
            "description": "A description of when the achievement should be awarded",
            "example": "Put on coffee for the gang."
          },
          "achievementPoints": {
            "type": "integer",
            "description": "How many achievement points that should be awarded when this achievement is achieved",
            "example": 500
          },
          "cooldownMinutes": {
            "type": "integer",
            "description": "Minutes until the this achievement becomes available again after awarding it",
            "example": 15,
            "default": 0
          }
        },
        "required": [
          "name",
          "description",
          "achievementPoints"
        ]
      },
      "Award": {
        "title": "Award",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "awardedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "achievementId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "userId",
          "achievementId"
        ]
      },
      "User": {
        "title": "User",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "example": "Niclas Lindstedt"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "niclas.lindstedt@etimo.se"
          },
          "slackHandle": {
            "type": "string",
            "description": "The @slack handle of the user",
            "example": "@niclas"
          }
        },
        "required": [
          "name",
          "email",
          "slackHandle"
        ]
      }
    }
  },
  "paths": {
    "/achievements": {
      "get": {
        "summary": "Get a list of achievements",
        "operationId": "getAchievements",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "The numbers of items to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Achievement"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Achievements"
        ]
      },
      "post": {
        "summary": "Create an achievement",
        "operationId": "createAchievement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Achievement"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "ID of the created resource."
                    }
                  }
                }
              }
            },
            "links": {
              "GetAchievementById": {
                "operationId": "getAchievement",
                "parameters": {
                  "achievementId": "$response.body#/id"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Achievements"
        ]
      }
    },
    "/achievements/{achievementId}": {
      "get": {
        "summary": "Get a single achievement",
        "operationId": "getAchievement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "achievementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Achievement"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          },
          "404": {
            "description": "The requested resource could not be found."
          }
        },
        "tags": [
          "Achievements"
        ]
      }
    },
    "/auth/login/{provider}": {
      "get": {
        "summary": "Start a login request",
        "operationId": "authLogin",
        "security": [],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "The OAuth provider to use for login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "301": {
            "description": "Redirection to OAuth2 service.",
            "headers": {
              "Location": {
                "description": "URI to OAuth2 token endpoint.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/callback/{provider}": {
      "get": {
        "summary": "Callback URL for OAuth2 code flow",
        "security": [],
        "operationId": "authCallback",
        "x-allow-unknown-query-parameters": true,
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "The OAuth provider to use for login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "The OAuth2 code to exchange for an access token.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "allowReserved": true
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/awards": {
      "get": {
        "summary": "Get a list of awards",
        "operationId": "getAwards",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "The numbers of items to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Award"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Awards"
        ]
      },
      "post": {
        "summary": "Give a user an award",
        "operationId": "createAward",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Award"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "ID of the created resource."
                    }
                  }
                }
              }
            },
            "links": {
              "GetAwardById": {
                "operationId": "getAward",
                "parameters": {
                  "userId": "$response.body#/id"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          },
          "404": {
            "description": "The user/achievement could not be found."
          }
        },
        "tags": [
          "Awards"
        ]
      }
    },
    "/awards/{awardId}": {
      "get": {
        "summary": "Get a single award",
        "operationId": "getAward",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "awardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Award"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          },
          "404": {
            "description": "The requested resource could not be found."
          }
        },
        "tags": [
          "Awards"
        ]
      }
    },
    "/slack/list-achievements": {
      "post": {
        "summary": "Display achievement list modal in Slack",
        "security": [
          {
            "ApiKeyParameter": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "The numbers of items to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The modal was created."
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Slack"
        ]
      }
    },
    "/slack/award-achievement": {
      "post": {
        "summary": "Show award achievement list modal in Slack",
        "security": [
          {
            "ApiKeyParameter": []
          }
        ],
        "responses": {
          "200": {
            "description": "The modal was created."
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Slack"
        ]
      }
    },
    "/slack/interact": {
      "post": {
        "summary": "Endpoint for slack to respond to interact messages (modals etc...)",
        "security": [
          {
            "ApiKeyParameter": []
          }
        ],
        "responses": {
          "200": {
            "description": "The modal was created."
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Slack"
        ]
      }
    },
    "/slack/create-achievement": {
      "post": {
        "summary": "Display achievement creation modal in Slack",
        "security": [
          {
            "ApiKeyParameter": []
          }
        ],
        "responses": {
          "200": {
            "description": "The modal was created."
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Slack"
        ]
      }
    },
    "/slack/sync-users": {
      "post": {
        "summary": "Sync slack users with the database",
        "security": [
          {
            "ApiKeyParameter": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users was synced."
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Slack"
        ]
      }
    },
    "/users": {
      "get": {
        "summary": "Get a list of users",
        "operationId": "getUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "The numbers of items to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "summary": "Create a user",
        "operationId": "createUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "ID of the created resource."
                    }
                  }
                }
              }
            },
            "links": {
              "GetUserById": {
                "operationId": "getUser",
                "parameters": {
                  "userId": "$response.body#/id"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "summary": "Get a single user",
        "operationId": "getUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid and could/would not be processed."
          },
          "401": {
            "description": "The request was unauthorized."
          },
          "404": {
            "description": "The requested resource could not be found."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/version": {
      "get": {
        "summary": "Get application version information",
        "security": [],
        "responses": {
          "200": {
            "description": "The request was successful"
          }
        },
        "tags": [
          "Version"
        ]
      }
    }
  },
  "tags": []
}